[eclipse/xtext#1629] moved xbase workflow to not to prefer xtend
