+++ org.eclipse.xtext.xbase.ui/src/org/eclipse/xtext/xbase/annotations/ui/XbaseWithAnnotationsUiModule.java
@@ -1,0 +1,16 @@
+/*
+ * generated by Xtext
+ */
+package org.eclipse.xtext.xbase.annotations.ui;
+
+import org.eclipse.ui.plugin.AbstractUIPlugin;
+
+/** 
+ * Use this class to register components to be used within the IDE.
+ */
+public class XbaseWithAnnotationsUiModule extends org.eclipse.xtext.xbase.annotations.ui.AbstractXbaseWithAnnotationsUiModule {
+
+	public XbaseWithAnnotationsUiModule(AbstractUIPlugin plugin) {
+		super(plugin);
+	}
+}
