--- org.eclipse.xtext.xbase.ui/src/org/eclipse/xtext/xbase/annotations/ui/XbaseWithAnnotationsUiModule.xtend
@@ -1,16 +1,0 @@
-/*
- * generated by Xtext
- */
-package org.eclipse.xtext.xbase.annotations.ui
-
-import org.eclipse.ui.plugin.AbstractUIPlugin
-
-/** 
- * Use this class to register components to be used within the IDE.
- */
-class XbaseWithAnnotationsUiModule extends org.eclipse.xtext.xbase.annotations.ui.AbstractXbaseWithAnnotationsUiModule {
-
-	new(AbstractUIPlugin plugin) {
-		super(plugin)
-	}
-}
