+++ org.eclipse.xtext.xbase.ui/src/org/eclipse/xtext/xbase/ui/XtypeUiModule.java
@@ -1,0 +1,16 @@
+/*
+ * generated by Xtext
+ */
+package org.eclipse.xtext.xbase.ui;
+
+import org.eclipse.ui.plugin.AbstractUIPlugin;
+
+/** 
+ * Use this class to register components to be used within the IDE.
+ */
+public class XtypeUiModule extends org.eclipse.xtext.xbase.ui.AbstractXtypeUiModule {
+
+	public XtypeUiModule(AbstractUIPlugin plugin) {
+		super(plugin);
+	}
+}
