--- org.eclipse.xtext.xbase.ui/src/org/eclipse/xtext/xbase/ui/XtypeUiModule.xtend
@@ -1,16 +1,0 @@
-/*
- * generated by Xtext
- */
-package org.eclipse.xtext.xbase.ui
-
-import org.eclipse.ui.plugin.AbstractUIPlugin
-
-/** 
- * Use this class to register components to be used within the IDE.
- */
-class XtypeUiModule extends org.eclipse.xtext.xbase.ui.AbstractXtypeUiModule {
-
-	new(AbstractUIPlugin plugin) {
-		super(plugin)
-	}
-}
