--- org.eclipse.xtext.xbase.ui/xtend-gen/org/eclipse/xtext/xbase/annotations/ui/XbaseWithAnnotationsUiModule.java
@@ -1,17 +1,0 @@
-/**
- * generated by Xtext
- */
-package org.eclipse.xtext.xbase.annotations.ui;
-
-import org.eclipse.ui.plugin.AbstractUIPlugin;
-import org.eclipse.xtext.xbase.annotations.ui.AbstractXbaseWithAnnotationsUiModule;
-
-/**
- * Use this class to register components to be used within the IDE.
- */
-@SuppressWarnings("all")
-public class XbaseWithAnnotationsUiModule extends AbstractXbaseWithAnnotationsUiModule {
-  public XbaseWithAnnotationsUiModule(final AbstractUIPlugin plugin) {
-    super(plugin);
-  }
-}
