--- org.eclipse.xtext.xbase.ui/xtend-gen/org/eclipse/xtext/xbase/ui/XtypeUiModule.java
@@ -1,17 +1,0 @@
-/**
- * generated by Xtext
- */
-package org.eclipse.xtext.xbase.ui;
-
-import org.eclipse.ui.plugin.AbstractUIPlugin;
-import org.eclipse.xtext.xbase.ui.AbstractXtypeUiModule;
-
-/**
- * Use this class to register components to be used within the IDE.
- */
-@SuppressWarnings("all")
-public class XtypeUiModule extends AbstractXtypeUiModule {
-  public XtypeUiModule(final AbstractUIPlugin plugin) {
-    super(plugin);
-  }
-}
